requirements.txt
psycopg2-binary == 2.9.5
pytest == 7.2.1
